Address PR #32 review: cap-proof stdin, atomic root/seal writes, path remotes
Confirmed bugs (each pinned by a new test):
- HookCommand: Console.In reads synchronously beneath its async surface,
  so the 3 s cap never bounded the stdin phase — park the read on a
  worker thread and WaitAsync on the cap; cap injectable for tests
- ProjectResolver: appending a root through a tracked save could clobber
  a concurrent hook's append (lost update) — append via guarded
  array_append in SQL and reload the tracked row
- CaptureService: seal was check-then-act — first-seal-wins is now the
  WHERE sealed_at IS NULL guard on an atomic update
- CaptureOptions: head+tail validated with long arithmetic so
  near-int.MaxValue knobs cannot wrap past the boot check
- RemoteIdentity: drive-letter and UNC remotes are paths, not scp hosts;
  C:\ and C:/ spellings of one directory are now one identity

Cleanups from the same review:
- ProjectLocator: toplevel and remote git lookups start concurrently
  (one round-trip on the prompt hook's 500 ms budget); stderr drain task
  observed, not discarded
- DbRaces: retries now fire only on unique violations, with named
  attempt caps explaining the 5-vs-3 split
- PayloadTruncator: original size read from the raw UTF-8 view, no
  UTF-16 round-trip
- TestPostgres: one home for the test connection string + skip message
- Program.cs: HttpClient.Timeout documented as a deliberate backstop

Declined with rationale (recorded in docstrings): lowercasing owner/repo
case (irreversible-merge risk on case-sensitive hosts), combined OR
project query (breaks identity-first precedence), event-map refactor
(the §3 enum is closed).

+++ src/Mimir.Server/Configuration/CaptureOptions.cs
@@ -5,6 +5,10 @@ namespace Mimir.Server.Configuration;
 /// <summary>
 /// Spec §11: event payload cap — 4 KB per payload field, kept as 3 KB head + 1 KB tail. The loss
 /// is deliberate (ADR-0003): capture is lossy-by-design and the original size is always recorded.
+/// The cap and head/tail are independent knobs: a field at or under the cap is stored whole, an
+/// oversized one collapses to head + tail — so a cap raised well above head + tail widens the
+/// size drop at the boundary. The cap decides when to truncate; head and tail decide what
+/// survives.
 /// </summary>
 public sealed class CaptureOptions : IValidatableObject
 {
@@ -24,7 +28,8 @@ public sealed class CaptureOptions : IValidatableObject
 
     public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
     {
-        if (PayloadHeadBytes + PayloadTailBytes > PayloadFieldCapBytes)
+        // Summed as long: two near-int.MaxValue knobs must not wrap around the check.
+        if ((long)PayloadHeadBytes + PayloadTailBytes > PayloadFieldCapBytes)
         {
             yield return new ValidationResult(
                 $"PayloadHeadBytes ({PayloadHeadBytes}) + PayloadTailBytes ({PayloadTailBytes}) must not "
